[22/11/09] BOJ #1012 - DFS
